requirements.txt
bokeh>= 2.2
colorcet
firecloud_dalmatian>=0.0.17
gseapy==0.9.18
gsheets==0.4.1
gspread==3.6.0
matplotlib
oauth2client>=4.1.3
pandas
pybedtools
pyBigWig
pysam
pytest
requests>=2.24.0
scikit_learn
scipy>=1.0.0
seaborn
statsmodels
taigapy>=2.12
venn
biomart